Move evaluators to their own package
